pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "atsphinx-og-article"
authors = [{name = "Kazuya Takei", email = "myself@attakei.net"}]
license = {file = "LICENSE"}
classifiers = [
  "Development Status :: 3 - Alpha",
  "Framework :: Sphinx",
  "Framework :: Sphinx :: Extension",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Documentation :: Sphinx",
]
readme = "README.rst"
dependencies = [
  "python-dateutil",
  "sphinx",
]
dynamic = ["version", "description"]

[project.optional-dependencies]
test = [
  "pytest ==7.*",
]

[project.urls]
Home = "https://github.com/atsphinx/og-article"

[tool.flit.module]
name = "atsphinx.og_article"

[tool.flit.sdist]
exclude = [".github", "tests"]
